initial UI for currency selection
