+++ frontend/composables/use-preferences.ts
@@ -1,35 +1,5 @@
 import { Ref } from "vue";
-
-export type DaisyTheme =
-  | "light"
-  | "dark"
-  | "cupcake"
-  | "bumblebee"
-  | "emerald"
-  | "corporate"
-  | "synthwave"
-  | "retro"
-  | "cyberpunk"
-  | "valentine"
-  | "halloween"
-  | "garden"
-  | "forest"
-  | "aqua"
-  | "lofi"
-  | "pastel"
-  | "fantasy"
-  | "wireframe"
-  | "black"
-  | "luxury"
-  | "dracula"
-  | "cmyk"
-  | "autumn"
-  | "business"
-  | "acid"
-  | "lemonade"
-  | "night"
-  | "coffee"
-  | "winter";
+import { DaisyTheme } from "~~/lib/data/themes";
 
 export type LocationViewPreferences = {
   showDetails: boolean;
